Custom dropdown w/congressional targets for DACA event creation
If user lands on event create page as recognized user, AND
category "daca" is prefilled in query string, AND
event creation page has a custom extra_json_data field with
targeting information, THEN
create a new dropdown menu listing the user's senators and congressperson,
and use javascript to dynamically change the event title + description
with the appropriate representative name when the dropdown value is changed
by the user.

Dropdown value also ends up in a custom eventfield if we ever want it for structured reporting.